Add js2-mode settings snippet.
Add a code snippet to make tabnine-company work with js2-mode.
It is better for js2-mode users, and mostly with Doom Emacs users that use js2-mode as default javascript mode. The company code snippet in the README.md most of the times don't work with Doom Emacs users that use it default configuration for javascript.